browser/fuel/test/test_Browser.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>
<window title="Testing Browser"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
  xmlns:html="http://www.w3.org/1999/xhtml">
  
  <script type="application/javascript" src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>

  <body xmlns="http://www.w3.org/1999/xhtml" />

<script type="application/javascript">
<![CDATA[
const Ci = Components.interfaces;
const Cc = Components.classes;

test_Browser();

function url(spec) {
  var ios = Cc["@mozilla.org/network/io-service;1"].getService(Ci.nsIIOService);
  return ios.newURI(spec, null, null);
}

var gTabOpenCount = 0;
var gTabCloseCount = 0;
var gTabMoveCount = 0;

function test_Browser() {
  var windows = Application.windows;
  ok(windows, "Check access to browser windows");
  ok(windows.length, "There should be at least one browser window open");
  
  var activeWin = Application.activeWindow;
  activeWin.events.addListener("TabOpen", onTabOpen);
  activeWin.events.addListener("TabClose", onTabClose);
  activeWin.events.addListener("TabMove", onTabMove);
  
  var pageA = activeWin.open(url("chrome://mochikit/content/chrome/browser/fuel/test/test_ContentA.html"));
  var pageB = activeWin.open(url("chrome://mochikit/content/chrome/browser/fuel/test/test_ContentB.html"));
  pageB.focus();

  is(activeWin.tabs.length, 3, "Checking length of 'Browser.tabs' after opening 2 additional tabs");
  is(activeWin.activeTab.index, pageB.index, "Checking 'Browser.activeTab' after setting focus");
  
  SimpleTest.waitForExplicitFinish();
  setTimeout(afterOpen, 1000);
  
  // need to wait for the url's to be refreshed during the load
  function afterOpen() {
    is(pageA.uri.spec, "chrome://mochikit/content/chrome/browser/fuel/test/test_ContentA.html", "Checking 'BrowserTab.uri' after opening");  
    is(pageB.uri.spec, "chrome://mochikit/content/chrome/browser/fuel/test/test_ContentB.html", "Checking 'BrowserTab.uri' after opening");  

    // check event
    todo_is(gTabOpenCount, 2, "Checking event handler for tab open");
    
    // test document access
    var test1 = pageA.document.getElementById("test1");
    ok(test1, "Checking existence of element in content DOM");
    is(test1.innerHTML, "A", "Checking content of element in content DOM");
    
    // test moving tab
    pageA.moveToEnd();
    is(pageA.index, 2, "Checking index after moving tab");
    
    // check event
    is(gTabMoveCount, 1, "Checking event handler for tab move");

    // test loading new content into a tab
    // the event will be checked in afterClose
    pageA.events.addListener("load", onPageLoad);
    pageA.load(pageB.uri);
  }
  
  function onPageLoad(event) {
    is(pageA.uri.spec, "chrome://mochikit/content/chrome/browser/fuel/test/test_ContentB.html", "Checking 'BrowserTab.uri' after loading new content");  

    // start testing closing tabs
    pageA.close();
    pageB.close();
    setTimeout(afterClose, 1000);
  }

  function afterClose() {
    // check event
    is(gTabCloseCount, 2, "Checking event handler for tab close");
    
    is(activeWin.tabs.length, 1, "Checking length of 'Browser.tabs' after closing 2 tabs");
    
    SimpleTest.finish();
  }
}

function onTabOpen(event) {
  gTabOpenCount++;
}

function onTabClose(event) {
  gTabCloseCount++;
}

function onTabMove(event) {
  gTabMoveCount++;
}

]]>
</script>

</window>
